+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Added:
 - Added `ONTraC_analysis` and `analysis` module
 - Added `niche cluster` part in `GP.py`
 - Added version information at the top of the output
+- Added flexible cell type coding option
 
 Changed: